refactor: drawn 컴포넌트 삭제 후 기존 페이지에 병합
