+++ src/home/pages/Withdraw/Withdraw.tsx
@@ -10,12 +10,25 @@ import {
   StyledListItem,
 } from './Withdraw.style';
 import { BoxButton, CheckBox } from '@yourssu/design-system-react';
-import { useState } from 'react';
-import { Withdrawn } from '@/home/components/WithdrawContents/Withdrawn/Withdrawn';
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 export const Withdraw = () => {
   const [agreed, setAgreed] = useState(false);
   const [draw, setDraw] = useState(false);
+
+  const Withdrawn = () => {
+    const navigate = useNavigate();
+
+    useEffect(() => {
+      const timer = setTimeout(() => {
+        navigate('/');
+      }, 3000);
+      return () => clearTimeout(timer);
+    }, [navigate]);
+
+    return <p>계정 탈퇴가 완료되었습니다.</p>;
+  };
 
   const handleCheckAgree = () => {
     setAgreed(!agreed);
